+++ src/Providers/Editor/EditorControllerProvider.cs
@@ -5,17 +5,18 @@ using System.Reflection;
 namespace DialToolsForVS
 {
     [Export(typeof(IDialControllerProvider))]
-    internal class ScrollControllerProvider : IDialControllerProvider
+    internal class EditorControllerProvider : IDialControllerProvider
     {
-        public const string Moniker = "Scroll";
+        public const string Moniker = "Editor";
+
         public IDialController TryCreateController(IDialControllerHost host)
         {
             string folder = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            string iconFilePath = Path.Combine(folder, "Providers\\Scroll\\icon.png");
+            string iconFilePath = Path.Combine(folder, "Providers\\Editor\\icon.png");
 
             host.AddMenuItem(Moniker, iconFilePath);
 
-            return new ScrollController();
+            return new EditorController();
         }
     }
 }