+++ src/Providers/Navigate/NavigateController.cs
@@ -2,7 +2,7 @@
 {
     internal class NavigateController : IDialController
     {
-        public string Moniker => PredefinedMonikers.Navigate;
+        public string Moniker => NavigateControllerProvider.Moniker;
         public Specificity Specificity => Specificity.Global;
 
         public bool CanHandleClick => false;