+++ src/Providers/Navigate/NavigateControllerProvider.cs
@@ -8,12 +8,13 @@ namespace DialToolsForVS
     [Export(typeof(IDialControllerProvider))]
     internal class NavigateControllerProvider : IDialControllerProvider
     {
+        public const string Moniker = "Navigation";
         public IDialController TryCreateController(IDialControllerHost host)
         {
             string folder = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
             string iconFilePath = Path.Combine(folder, "Providers\\Navigate\\icon.png");
 
-            host.AddMenuItem(PredefinedMonikers.Navigate, iconFilePath);
+            host.AddMenuItem(Moniker, iconFilePath);
 
             return new NavigateController();
         }