+++ src/Providers/Scroll/ScrollController.cs
@@ -6,7 +6,7 @@ namespace DialToolsForVS
 {
     internal class ScrollController : IDialController
     {
-        public string Moniker => PredefinedMonikers.Scroll;
+        public string Moniker => ScrollControllerProvider.Moniker;
         public Specificity Specificity => (Specificity)int.MaxValue;
         public bool CanHandleClick => true;
         public bool CanHandleRotate => true;