+++ src/Providers/Zoom/ZoomControllerProvider.cs
@@ -8,12 +8,13 @@ namespace DialToolsForVS
     [Export(typeof(IDialControllerProvider))]
     internal class ZoomControllerProvider : IDialControllerProvider
     {
+        public const string Moniker = "Zoom";
         public IDialController TryCreateController(IDialControllerHost host)
         {
             string folder = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
             string iconFilePath = Path.Combine(folder, "Providers\\Zoom\\icon.png");
 
-            host.AddMenuItem(PredefinedMonikers.Zoom, iconFilePath);
+            host.AddMenuItem(Moniker, iconFilePath);
 
             return new ZoomController();
         }