+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [x] Add debug navigation
 - [x] Reset zoom level on click
 - [x] Add status bar text to show selected menu item
+- [x] Add editor shifter
 - [ ] Add telemtry
 
 Features that have a checkmark are complete and available for